requirements.txt
azure-functions
openai
requests
azure-identity
azure-keyvault-secrets
azure-cosmos
httpx